[1701]: Split out the command setup funcs into query_setup.go and tx_setup.go. Reorg the stuff in flags.go and helpers.go to hopefully make more sense.
