+++ x/exchange/client/cli/helpers.go
@@ -2,10 +2,7 @@ package cli
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"regexp"
-	"strconv"
 	"strings"
 
 	"github.com/gogo/protobuf/proto"
@@ -34,11 +31,11 @@ var (
 )
 
 // A msgMaker is a function that makes a Msg from a client.Context, FlagSet, and set of args.
-type msgMaker func(clientCtx client.Context, flagSet *pflag.FlagSet, args []string) (sdk.Msg, error)
+type msgMaker[R sdk.Msg] func(clientCtx client.Context, flagSet *pflag.FlagSet, args []string) (R, error)
 
-// genericTxRunE returns a cobra.Command.RunE function that gets the client.Context, and FlagSet,
+// genericTxRunE returns a cobra.Command.RunE function that gets the client.Context and FlagSet,
 // then uses the provided maker to make the Msg that it then generates or broadcasts as a Tx.
-func genericTxRunE(maker msgMaker) func(cmd *cobra.Command, args []string) error {
+func genericTxRunE[R sdk.Msg](maker msgMaker[R]) func(cmd *cobra.Command, args []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		clientCtx, err := client.GetClientTxContext(cmd)
 		if err != nil {
@@ -56,10 +53,10 @@ func genericTxRunE(maker msgMaker) func(cmd *cobra.Command, args []string) error
 	}
 }
 
-// govTxRunE returns a cobra.Command.RunE function that gets the client.Context, and FlagSet,
-// then uses the provided maker to make the Msg. The Msg is then either generated or put in a
-// governance proposal which is then broadcast as a Tx.
-func govTxRunE(maker msgMaker) func(cmd *cobra.Command, args []string) error {
+// govTxRunE returns a cobra.Command.RunE function that gets the client.Context and FlagSet,
+// then uses the provided maker to make the Msg. The Msg is then put into a governance
+// proposal and either generated or broadcast as a Tx.
+func govTxRunE[R sdk.Msg](maker msgMaker[R]) func(cmd *cobra.Command, args []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		clientCtx, err := client.GetClientTxContext(cmd)
 		if err != nil {
@@ -121,227 +118,6 @@ func MarkFlagsRequired(cmd *cobra.Command, names ...string) {
 			panic(fmt.Errorf("error marking --%s flag required on %s: %w", name, cmd.Name(), err))
 		}
 	}
-}
-
-// ReadCoinsFlag reads a string flag and converts it into sdk.Coins.
-// If the flag wasn't provided, this returns nil, nil.
-func ReadCoinsFlag(flagSet *pflag.FlagSet, name string) (sdk.Coins, error) {
-	value, err := flagSet.GetString(name)
-	if len(value) == 0 || err != nil {
-		return nil, err
-	}
-	rv, err := sdk.ParseCoinsNormalized(value)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing --%s value %q as coins: %w", name, value, err)
-	}
-	return rv, nil
-}
-
-// ReadCoinFlag reads a string flag and converts it into *sdk.Coin.
-// If the flag wasn't provided, this returns nil, nil.
-func ReadCoinFlag(flagSet *pflag.FlagSet, name string) (*sdk.Coin, error) {
-	value, err := flagSet.GetString(name)
-	if len(value) == 0 || err != nil {
-		return nil, err
-	}
-	rv, err := exchange.ParseCoin(value)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing --%s value %q as a coin: %w", name, value, err)
-	}
-	return &rv, nil
-}
-
-// ReadReqCoinFlag reads a string flag and converts it into a sdk.Coin and requires it to have a value.
-func ReadReqCoinFlag(flagSet *pflag.FlagSet, name string) (sdk.Coin, error) {
-	rv, err := ReadCoinFlag(flagSet, name)
-	if err != nil {
-		return sdk.Coin{}, err
-	}
-	if rv == nil {
-		return sdk.Coin{}, fmt.Errorf("missing required --%s flag", name)
-	}
-	return *rv, nil
-}
-
-// ReadOrderIdsFlag reads a UintSlice flag and converts it into a []uint64.
-func ReadOrderIdsFlag(flagSet *pflag.FlagSet, name string) ([]uint64, error) {
-	ids, err := flagSet.GetUintSlice(name)
-	if err != nil {
-		return nil, err
-	}
-	rv := make([]uint64, len(ids))
-	for i, id := range ids {
-		rv[i] = uint64(id)
-	}
-	return rv, nil
-}
-
-// ReadAddrFlagOrFrom gets the requested flag or, if it wasn't provided, gets the --from address.
-func ReadAddrFlagOrFrom(clientCtx client.Context, flagSet *pflag.FlagSet, name string) (string, error) {
-	rv, err := flagSet.GetString(name)
-	if len(rv) > 0 || err != nil {
-		return rv, err
-	}
-
-	rv = clientCtx.FromAddress.String()
-	if len(rv) > 0 {
-		return rv, nil
-	}
-
-	return "", fmt.Errorf("no %s provided", name)
-}
-
-// ReadAccessGrantsFlag reads a StringSlice flag and converts it to a slice of AccessGrants.
-func ReadAccessGrantsFlag(flagSet *pflag.FlagSet, name string) ([]exchange.AccessGrant, error) {
-	vals, err := flagSet.GetStringSlice(name)
-	if len(vals) == 0 || err != nil {
-		return nil, err
-	}
-	return ParseAccessGrants(vals)
-}
-
-// permSepRx is a regexp that matches characters that can be used to separate permissions.
-var permSepRx = regexp.MustCompile(`[ +-.]`)
-
-// ParseAccessGrant parses an AccessGrant from a string with the format "<address>:<perm 1>[+<perm 2>...]".
-func ParseAccessGrant(val string) (*exchange.AccessGrant, error) {
-	parts := strings.Split(val, ":")
-	if len(parts) != 2 {
-		return nil, fmt.Errorf("could not parse %q as an <access grant>: expected format <address>:<permissions>", val)
-	}
-
-	addr := strings.TrimSpace(parts[0])
-	perms := strings.ToLower(strings.TrimSpace(parts[1]))
-	if len(addr) == 0 || len(perms) == 0 {
-		return nil, fmt.Errorf("invalid <access grant>: both an <address> and <permissions> are required")
-	}
-
-	rv := &exchange.AccessGrant{Address: addr}
-
-	if perms == "all" {
-		rv.Permissions = exchange.AllPermissions()
-		return rv, nil
-	}
-
-	permVals := permSepRx.Split(perms, -1)
-	var err error
-	rv.Permissions, err = exchange.ParsePermissions(permVals...)
-	if err != nil {
-		return nil, fmt.Errorf("could not parse permissions for %q from %q: %w", rv.Address, parts[1], err)
-	}
-
-	return rv, nil
-}
-
-// ParseAccessGrants parses an AccessGrant from each of the provided vals.
-func ParseAccessGrants(vals []string) ([]exchange.AccessGrant, error) {
-	var errs []error
-	grants := make([]exchange.AccessGrant, 0, len(vals))
-	for _, val := range vals {
-		ag, err := ParseAccessGrant(val)
-		if err != nil {
-			errs = append(errs, err)
-		}
-		if ag != nil {
-			grants = append(grants, *ag)
-		}
-	}
-	return grants, errors.Join(errs...)
-}
-
-// ReadFlatFeeFlag reads a StringSlice flag and converts it into a slice of sdk.Coin.
-func ReadFlatFeeFlag(flagSet *pflag.FlagSet, name string) ([]sdk.Coin, error) {
-	vals, err := flagSet.GetStringSlice(name)
-	if len(vals) == 0 || err != nil {
-		return nil, err
-	}
-	return ParseFlatFeeOptions(vals)
-}
-
-// ParseFlatFeeOptions parses an sdk.Coin from each of the provided vals.
-func ParseFlatFeeOptions(vals []string) ([]sdk.Coin, error) {
-	var errs []error
-	rv := make([]sdk.Coin, 0, len(vals))
-	for _, val := range vals {
-		coin, err := exchange.ParseCoin(val)
-		if err != nil {
-			errs = append(errs, err)
-		} else {
-			rv = append(rv, coin)
-		}
-	}
-	return rv, errors.Join(errs...)
-}
-
-// ReadFeeRatiosFlag reads a StringSlice flag and converts it into a slice of exchange.FeeRatio.
-func ReadFeeRatiosFlag(flagSet *pflag.FlagSet, name string) ([]exchange.FeeRatio, error) {
-	vals, err := flagSet.GetStringSlice(name)
-	if len(vals) == 0 || err != nil {
-		return nil, err
-	}
-	return ParseFeeRatios(vals)
-}
-
-// ParseFeeRatios parses a FeeRatio from each of the provided vals.
-func ParseFeeRatios(vals []string) ([]exchange.FeeRatio, error) {
-	var errs []error
-	ratios := make([]exchange.FeeRatio, 0, len(vals))
-	for _, val := range vals {
-		ratio, err := exchange.ParseFeeRatio(val)
-		if err != nil {
-			errs = append(errs, err)
-		}
-		if ratio != nil {
-			ratios = append(ratios, *ratio)
-		}
-	}
-	return ratios, errors.Join(errs...)
-}
-
-// ReadFlagSplitsFlag reads a StringSlice flag and converts it into a slice of exchange.DenomSplit.
-func ReadFlagSplitsFlag(flagSet *pflag.FlagSet, name string) ([]exchange.DenomSplit, error) {
-	vals, err := flagSet.GetStringSlice(FlagSplit)
-	if len(vals) == 0 || err != nil {
-		return nil, err
-	}
-	return ParseSplits(vals)
-}
-
-// ParseSplit parses a DenomSplit from a string with the format "<denom>:<amount>".
-func ParseSplit(val string) (*exchange.DenomSplit, error) {
-	parts := strings.Split(val, ":")
-	if len(parts) != 2 {
-		return nil, fmt.Errorf("invalid denom split %q: expected format <denom>:<amount>")
-	}
-
-	denom := strings.TrimSpace(parts[0])
-	amountStr := strings.TrimSpace(parts[1])
-	if len(denom) == 0 || len(amountStr) == 0 {
-		return nil, fmt.Errorf("invalid denom split %q: both a <denom> and <amount> are required", val)
-	}
-
-	amount, err := strconv.ParseUint(amountStr, 10, 32)
-	if err != nil {
-		return nil, fmt.Errorf("could not parse %s amount %q into uint32: %w", denom, amountStr, err)
-	}
-
-	return &exchange.DenomSplit{Denom: denom, Split: uint32(amount)}, nil
-}
-
-// ParseSplits parses a DenomSplit from each of the provided vals.
-func ParseSplits(vals []string) ([]exchange.DenomSplit, error) {
-	var errs []error
-	splits := make([]exchange.DenomSplit, 0, len(vals))
-	for _, val := range vals {
-		split, err := ParseSplit(val)
-		if err != nil {
-			errs = append(errs, err)
-		}
-		if split != nil {
-			splits = append(splits, *split)
-		}
-	}
-	return splits, errors.Join(errs...)
 }
 
 // AddUseArgs adds the given args to the cmd's Use.
